+++ src/string/rule.js
@@ -40,6 +40,8 @@ function Rule (subrules, type, handler, options) {
   this.break = options._p_break
 
   this.bufferMode = (options._bufferMode === true)
+  this.groupStart = options._groupStart
+  this.groupEnd = options._groupEnd
 
   this.atok = options